added 'value' field to AccessToken
This is the JWT string representation of the access token. It will useful for scenarios where
we want to provide handle of actual string token so it can be used as a header to another HTTP server.

For example, when service A get's the token from client but also needs to make a call to Service B. Service B is also protected by auth and needs the token. Now that Service A has handle of string value of jwt access token, it can include it in the header of HTTP request to Service B.