#CSW-89: Bilal | Himanshu : Enabled auth for location service for public network
---
for `publicNetwork` http binding, auth is enabled for register and unregister apis. For local mode auth is disabled for all endpoints. By default `publicNetwork` is set to `false` and hence auth is not required. Since we have not yet decided on auth policies, we have used EmptyPolicy for now. This will require user to be logged in but will not require any roles.

Co-authored-by: Himanshu <hmertia@thoughtworks.com>